# FurnitureCompany/bin/Release/net6.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "FurnitureCompany, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
...
    "/employee": {
      "post": {
        "tags": [
          "Employee"
        ],
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/EmployeeDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/employee/employeeUpdateAvatar/{id}": {
      "put": {
        "tags": [
          "Employee"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "query",
            "name": "newAvatarUrl",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/employee/emplooyeeUpdateInformation/{id}": {
      "put": {
        "tags": [
          "Employee"
        ],
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/EmployeeDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/employee/updateEmployeeStatus/{id}": {
      "put": {
        "tags": [
          "Employee"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "query",
            "name": "workStatus",
            "type": "boolean"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/employee/ViewAssign/employeeId/{id}": {
      "get": {
        "tags": [
          "Employee"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/manager/updateOrderStatusDone/{id}": {
      "put": {
        "tags": [
          "Manager"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "orderId",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/manager/updateTotalPrice/order/{id}": {
      "put": {
        "tags": [
          "Manager"
        ],
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "orderId",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/OrderDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/role/{id}": {
      "get": {
        "tags": [
          "Role"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
    "/AddNewRole": {
      "post": {
        "tags": [
          "Role"
        ],
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/RoleDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/role/removerole/{id}": {
      "delete": {
        "tags": [
          "Role"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/service/serviceID/{id}": {
      "get": {
        "tags": [
          "Service"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/AddNewService": {
      "post": {
        "tags": [
          "Service"
        ],
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/ServiceDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/service/RemoveService/{id}": {
      "delete": {
        "tags": [
          "Service"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/UpdateService/{id}": {
      "put": {
        "tags": [
          "Service"
        ],
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/ServiceDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "EmployeeDto": {
      "type": "object",
      "properties": {
        "imageUrl": {
          "type": "string"
        },
        "accountId": {
          "format": "int32",
          "type": "integer"
        },
        "specialtyId": {
          "format": "int32",
          "type": "integer"
        },
        "employeeName": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "employeePhoneNumber": {
          "type": "string"
        },
        "status": {
          "type": "boolean"
        }
      }
    },
    "OrderDto": {
      "type": "object",
      "properties": {
        "orderId": {
          "format": "int32",
          "type": "integer"
        },
        "customerId": {
          "format": "int32",
          "type": "integer"
        },
        "workingStatusId": {
          "format": "int32",
          "type": "integer"
        },
        "address": {
          "type": "string"
        },
        "totalPrice": {
          "type": "string"
        },
        "createAt": {
          "format": "date-time",
          "type": "string"
        },
        "updateAt": {
          "format": "date-time",
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "status": {
          "type": "boolean"
        }
      }
    },
    "RoleDto": {
      "type": "object",
      "properties": {
        "roleName": {
          "type": "string"
        }
      }
    },
...
    "ServiceDto": {
      "type": "object",
      "properties": {
        "serviceName": {
          "type": "string"
        },
        "serviceDescription": {
          "type": "string"
        },
        "price": {
          "type": "string"
        },
        "type": {
          "format": "int32",
          "type": "integer"
        },
        "createAt": {
          "format": "date-time",
          "type": "string"
        },
        "updateAt": {
          "format": "date-time",
          "type": "string"
        },
        "status": {
          "type": "boolean"
        }
      }
    },
...
  }
}